+++ src/db/ColumnSchema.php
@@ -1,9 +1,6 @@
 <?php
-/**
- * @link https://www.yiiframework.com/
- * @copyright Copyright (c) 2008 Yii Software LLC
- * @license https://www.yiiframework.com/license/
- */
+
+declare(strict_types=1);
 
 namespace yii\db;
 
@@ -12,93 +9,105 @@ use yii\helpers\StringHelper;
 
 /**
  * ColumnSchema class describes the metadata of a column in a database table.
- *
- * @author Qiang Xue <qiang.xue@gmail.com>
- * @since 2.0
  */
 class ColumnSchema extends BaseObject
 {
     /**
-     * @var string name of this column (without quotes).
+     * @var string name of this column (without quotes)
      */
-    public $name;
+    public string $name = '';
     /**
-     * @var bool whether this column can be null.
+     * @var string abstract type of this column.
+     * Possible abstract types include:
+     * - `char`, `string`, `text`, `boolean`, `smallint`, `integer`, `bigint`, `float`, `decimal`, `datetime`,
+     *   `timestamp`, `time`, `date`, `binary`, `money`, `json`, `uuid`, and `uuid_pk`
      */
-    public $allowNull;
-    /**
-     * @var string abstract type of this column. Possible abstract types include:
-     * char, string, text, boolean, smallint, integer, bigint, float, decimal, datetime,
-     * timestamp, time, date, binary, and money.
-     */
-    public $type;
+    public string $type = '';
     /**
      * @var string the PHP type of this column. Possible PHP types include:
-     * `string`, `boolean`, `integer`, `double`, `array`.
+     * `PHPType::INTEGER`, `PHPType::STRING`, `PHPType::BOOLEAN`, `PHPType::DOUBLE`, `PHPType::RESOURCE`,
+     * `PHPType::ARRAY`, `PHPType::OBJECT`, and `PHPType::NULL`
+     *
+     * @see [[\yii\db\enum\PHPType]]
      */
-    public $phpType;
+    public string $phpType = 'string';
     /**
-     * @var string the DB type of this column. Possible DB types vary according to the type of DBMS.
+     * @var string the DB type of this column. Possible DB types vary according to the type of DBMS
      */
-    public $dbType;
-    /**
-     * @var mixed default value of this column
-     */
-    public $defaultValue;
-    /**
-     * @var array enumerable values. This is set only if the column is declared to be an enumerable type.
-     */
-    public $enumValues;
-    /**
-     * @var int display size of the column.
-     */
-    public $size;
-    /**
-     * @var int precision of the column data, if it is numeric.
-     */
-    public $precision;
-    /**
-     * @var int scale of the column data, if it is numeric.
-     */
-    public $scale;
+    public string $dbType = '';
     /**
      * @var bool whether this column is a primary key
      */
-    public $isPrimaryKey;
+    public bool $isPrimaryKey = false;
     /**
      * @var bool whether this column is auto-incremental
      */
-    public $autoIncrement = false;
+    public bool $autoIncrement = false;
     /**
-     * @var bool whether this column is unsigned. This is only meaningful
-     * when [[type]] is `smallint`, `integer` or `bigint`.
+     * @var bool whether this column is unique constraint
      */
-    public $unsigned;
+    public bool $isUnique = false;
     /**
-     * @var string comment of this column. Not all DBMS support this.
+     * @var string|null whether this column is check constraint
      */
-    public $comment;
-
+    public string|null $checkExpression = null;
+    /**
+     * @var mixed default value of this column
+     */
+    public mixed $defaultValue = null;
+    /**
+     * @var bool whether this column can be null
+     */
+    public bool $allowNull = false;
+    /**
+     * @var int|null display size of the column
+     */
+    public int|null $size = null;
+    /**
+     * @var int|null precision of the column data, if it is numeric
+     */
+    public $precision;
+    /**
+     * @var int|null scale of the column data, if it is numeric
+     */
+    public int|null $scale = null;
+    /**
+     * @var bool whether this column is unsigned
+     * This is only meaningful when [[type]] is `smallint`, `integer` or `bigint`
+     */
+    public bool $unsigned = false;
+    /**
+     * @var array enumerable values. This is set only if the column is declared to be an enumerable type
+     */
+    public array $enumValues = [];
+    /**
+     * @var string|null comment of this column. Not all DBMS support this
+     */
+    public string|null $comment = null;
 
     /**
      * Converts the input value according to [[phpType]] after retrieval from the database.
-     * If the value is null or an [[Expression]], it will not be converted.
-     * @param mixed $value input value
-     * @return mixed converted value
+     * If the value is `null` or an [[Expression]], it will not be converted.
+     *
+     * @param mixed $value input value.
+     *
+     * @return mixed converted value.
      */
-    public function phpTypecast($value)
+    public function phpTypecast(mixed $value): mixed
     {
         return $this->typecast($value);
     }
 
     /**
      * Converts the input value according to [[type]] and [[dbType]] for use in a db query.
-     * If the value is null or an [[Expression]], it will not be converted.
-     * @param mixed $value input value
-     * @return mixed converted value. This may also be an array containing the value as the first element
-     * and the PDO type as the second element.
+     * If the value is `null` or an [[Expression]], it will not be converted.
+     *
+     * @param mixed $value input value.
+     *
+     * @return mixed converted value. This may also be an array containing the value as the first element and the PDO
+     * type as the second element.
      */
-    public function dbTypecast($value)
+    public function dbTypecast(mixed $value): mixed
     {
         // the default implementation does the same as casting for PHP, but it should be possible
         // to override this with annotation of explicit PDO type.
@@ -107,91 +116,177 @@ class ColumnSchema extends BaseObject
 
     /**
      * Converts the input value according to [[phpType]] after retrieval from the database.
-     * If the value is null or an [[Expression]], it will not be converted.
-     * @param mixed $value input value
-     * @return mixed converted value
-     * @since 2.0.3
+     * If the value is `null` or an [[Expression]], it will not be converted.
+     *
+     * @param mixed $value input value.
+     *
+     * @return mixed converted value.
      */
-    protected function typecast($value)
+    protected function typecast(mixed $value): mixed
     {
-        if (
-            $value === ''
-            && !in_array(
-                $this->type,
-                [
-                    Schema::TYPE_TEXT,
-                    Schema::TYPE_STRING,
-                    Schema::TYPE_BINARY,
-                    Schema::TYPE_CHAR
-                ],
-                true
-            )
-        ) {
+        if ($value === null || $this->isEmptyValueForNonTextType($value)) {
             return null;
         }
 
-        if (
-            $value === null
-            || gettype($value) === $this->phpType
-            || $value instanceof ExpressionInterface
-            || $value instanceof Query
-        ) {
+        if ($value instanceof ExpressionInterface) {
             return $value;
         }
 
-        if (
-            is_array($value)
-            && count($value) === 2
-            && isset($value[1])
-            && in_array($value[1], $this->getPdoParamTypes(), true)
-        ) {
-            return new PdoValue($value[0], $value[1]);
-        }
+        return match ($this->phpType) {
+            'array' => $this->typeCastAsArray($value),
+            'boolean' => $this->typeCastAsBoolean($value),
+            'double' => $this->typeCastAsDouble($value),
+            'integer' => $this->typeCastAsInteger($value),
+            'resource' => $this->typeCastAsResource($value),
+            'string' => $this->typeCastAsString($value),
+            default => $value,
+        };
+    }
 
-        switch ($this->phpType) {
-            case 'resource':
-            case 'string':
-                if (is_resource($value)) {
-                    return $value;
-                }
-                if (is_float($value)) {
-                    // ensure type cast always has . as decimal separator in all locales
-                    return StringHelper::floatToString($value);
-                }
-                if (
-                    is_numeric($value)
-                    && ColumnSchemaBuilder::CATEGORY_NUMERIC === ColumnSchemaBuilder::$typeCategoryMap[$this->type]
-                ) {
-                    // https://github.com/yiisoft/yii2/issues/14663
-                    return $value;
-                }
+    /**
+     * Checks if the given value is an empty string and the column type is not a text type.
+     *
+     * This method is used to determine if an empty string value should be treated as `null` for non-text type columns.
+     *
+     * @param mixed $value the value to check.
+     *
+     * @return bool `true` if the value is an empty string and the column is not a text type, false otherwise.
+     */
+    protected function isEmptyValueForNonTextType(mixed $value): bool
+    {
+        $textTypes = [
+            Schema::TYPE_BINARY,
+            Schema::TYPE_CHAR,
+            Schema::TYPE_STRING,
+            Schema::TYPE_TEXT,
+        ];
 
-                if (PHP_VERSION_ID >= 80100 && is_object($value) && $value instanceof \BackedEnum) {
-                    return (string) $value->value;
-                }
+        return ($value === '' && (is_string($value) && trim($value) === ''))
+            && !in_array($this->type, $textTypes, true);
+    }
 
-                return (string) $value;
-            case 'integer':
-                if (PHP_VERSION_ID >= 80100 && is_object($value) && $value instanceof \BackedEnum) {
-                    return (int) $value->value;
-                }
-                return (int) $value;
-            case 'boolean':
-                // treating a 0 bit value as false too
-                // https://github.com/yiisoft/yii2/issues/9006
-                return (bool) $value && $value !== "\0" && strtolower($value) !== 'false';
-            case 'double':
-                return (float) $value;
-        }
-
+    /**
+     * Converts a given database value to an `array` type.
+     *
+     * @param mixed $value the value to be cast. Can be a `string`, `number`, `boolean`, or `array`.
+     *
+     * @return mixed the result of the type cast.
+     */
+    protected function typeCastAsArray(mixed $value): mixed
+    {
         return $value;
     }
 
     /**
-     * @return int[] array of numbers that represent possible PDO parameter types
+     * Converts a given database value to a `boolean` type.
+     *
+     * @param mixed $value the value to be cast. Can be a `string`, `number`, `boolean`.
+     *
+     * @return bool the result of the type cast:
+     * - bool: `false` for strings `'false'`, `'0'`, or any value that equates to false, `true` for strings `'1'`, or
+     *   any non-empty, non-null value that equates to true.
      */
-    private function getPdoParamTypes()
+    protected function typeCastAsBoolean(mixed $value): bool
     {
-        return [\PDO::PARAM_BOOL, \PDO::PARAM_INT, \PDO::PARAM_STR, \PDO::PARAM_LOB, \PDO::PARAM_NULL, \PDO::PARAM_STMT];
+        // treating a 0 bit value as false too
+        // https://github.com/yiisoft/yii2/issues/9006
+        return (bool) $value && $value !== "\0" && strtolower((string) $value) !== 'false';
+    }
+
+    /**
+     * Converts a given database value to a `double` type.
+     *
+     * @param mixed $value the value to be cast. Can be a `string`, `number`.
+     *
+     * @return float the result of the type cast:
+     * - float: the float representation of a numeric value for numeric column types.
+     */
+    protected function typeCastAsDouble(mixed $value): float
+    {
+        return (float) $value;
+    }
+
+    /**
+     * Converts a given database value to an `integer` type.
+     *
+     * @param mixed $value the value to be cast. Can be a `string`, `number`, or a `BackedEnum` instance.
+     *
+     * @return int the result of the type cast:
+     * - [[BackedEnum]]: the `integer` value of the enum.
+     * - int: the integer representation of a numeric value for numeric column types.
+     */
+    protected function typeCastAsInteger(mixed $value): int|null
+    {
+        if ($value instanceof \BackedEnum) {
+            return (int) $value->value;
+        }
+
+        return (int) $value;
+    }
+
+    /**
+     * Converts a given database value to a `resource` type.
+     *
+     * @param mixed $value the value to be cast. Can be a `resource`.
+     *
+     * @return mixed the result of the type cast.
+     */
+    protected function typeCastAsResource(mixed $value): mixed
+    {
+        return $value;
+    }
+
+    /**
+     * Converts a given database value to a `string` type.
+     *
+     * @param mixed $value the value to be cast. Can be a `string`, `number` or `boolean`.
+     *
+     * @return int|string the result of the type cast:
+     * - string: the string representation of a float with '.' as decimal separator.
+     * - int: the integer representation of a numeric value for numeric column types.
+     * - string: the string value of a BackedEnum instance.
+     * - string: the string representation of any other value type.
+     */
+    protected function typeCastAsString(mixed $value): int|string
+    {
+        if (is_bool($value)) {
+            return $value ? '1' : '0';
+        }
+
+        if (is_float($value)) {
+            // ensure type cast always has . as decimal separator in all locales
+            return StringHelper::floatToString($value);
+        }
+
+        if (
+            is_numeric($value) &&
+            ColumnSchemaBuilder::CATEGORY_NUMERIC === ColumnSchemaBuilder::$typeCategoryMap[$this->type]
+        ) {
+            // https://github.com/yiisoft/yii2/issues/14663
+            return $value;
+        }
+
+        if ($value instanceof \BackedEnum) {
+            return (string) $value->value;
+        }
+
+        return (string) $value;
+    }
+
+    /**
+     * @return array array of numbers that represent possible PDO parameter types.
+     *
+     * @psalm-return int[] array of numbers that represent possible PDO parameter types.
+     */
+    private function getPdoParamTypes(): array
+    {
+        return [
+            \PDO::PARAM_BOOL,
+            \PDO::PARAM_INT,
+            \PDO::PARAM_STR,
+            \PDO::PARAM_LOB,
+            \PDO::PARAM_NULL,
+            \PDO::PARAM_STMT
+        ];
     }
 }