+++ src/db/mssql/ColumnSchema.php
@@ -19,7 +19,7 @@ class ColumnSchema extends \yii\db\ColumnSchema
     /**
      * @var bool whether this column is a computed column.
      */
-    public $isComputed;
+    public bool $isComputed = false;
 
     /**
      * Prepares default value and converts it according to [[phpType]].
@@ -28,7 +28,7 @@ class ColumnSchema extends \yii\db\ColumnSchema
      *
      * @return mixed converted value.
      */
-    public function defaultPhpTypecast($value)
+    public function defaultPhpTypecast(mixed $value): mixed
     {
         if ($value !== null) {
             // convert from MSSQL column_default format, e.g. ('1') -> 1, ('string') -> string
@@ -38,18 +38,38 @@ class ColumnSchema extends \yii\db\ColumnSchema
         return parent::phpTypecast($value);
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function dbTypecast(mixed $value): mixed
     {
-        if ($this->type === Schema::TYPE_BINARY && $this->dbType === 'varbinary') {
-            if ($value instanceof PdoValue && is_string($value->getValue())) {
-                $value = (string) $value->getValue();
-            }
-
-            if (is_string($value)) {
-                return new Expression('CONVERT(VARBINARY(MAX), ' . ('0x' . bin2hex($value)) . ')');
-            }
+        if ($this->dbType === 'varbinary') {
+            return $this->dbTypeCastAsVarbinary($value);
         }
 
         return parent::dbTypecast($value);
+    }
+
+    /**
+     * Typecasts a value to a `VARBINARY` database type.
+     *
+     * If the value is a string, it converts it into a hexadecimal representation and wraps it in an SQL expression for
+     * converting to `VARBINARY(MAX)`.
+     *
+     * @param mixed $value the value to be typecast.
+     *
+     * @return Expression the SQL expression representing the binary value in `VARBINARY` format.
+     */
+    protected function dbTypeCastAsVarbinary(mixed $value): mixed
+    {
+        if ($value instanceof PdoValue && is_string($value->getValue())) {
+            $value = (string) $value->getValue();
+        }
+
+        if (is_string($value)) {
+            return new Expression('CONVERT(VARBINARY(MAX), ' . ('0x' . bin2hex($value)) . ')');
+        }
+
+        return $value;
     }
 }